Set UnstakingDelay = 670 as a Global
Set UnstakingDelay = 670 as a Global

Removing this as an option on messages removes the risk of it being manually overriden(or set to 1). 

Without this, anyone can continuously stake, vote, unstake, and then send to another address and repeat the process. this leaves voting as insecure and is a loophole in the security model.